+++ src/OpenRTMPlugin/OpenRTMPlugin.cpp
@@ -256,7 +256,7 @@ public:
         connectionToSigSimulaionAboutToStart.disconnect();
         
         std::vector<RTC::RTObject_impl*> rtcs = manager->getComponents();
-        for(int i=0; i < rtcs.size(); ++i){
+        for(size_t i=0; i < rtcs.size(); ++i){
             RTC::RTObject_impl* rtc = rtcs[i];
             RTC::ExecutionContextList_var eclist = rtc->get_participating_contexts();
             if(eclist->length() > 0){
@@ -309,7 +309,7 @@ CNOID_EXPORT int cnoid::numUnmanagedRTCs()
 {
     int n = 0;
     std::vector<RTC::RTObject_impl*> rtcs = manager->getComponents();
-    for(int i=0; i < rtcs.size(); ++i){
+    for(size_t i=0; i < rtcs.size(); ++i){
         RTC::RTObject_impl* rtc = rtcs[i];
         if(managedComponents.find(rtc) == managedComponents.end()){
             ++n;
@@ -325,7 +325,7 @@ CNOID_EXPORT int cnoid::deleteUnmanagedRTCs()
     
     std::vector<RTC::RTObject_impl*> rtcs = manager->getComponents();
     //  deactivate
-    for(int i=0; i < rtcs.size(); ++i){
+    for(size_t i=0; i < rtcs.size(); ++i){
         RTC::RTObject_impl* rtc = rtcs[i];
         if(managedComponents.find(rtc) == managedComponents.end()){
         	RTC::ExecutionContextList_var eclist = rtc->get_owned_contexts();
@@ -346,7 +346,7 @@ CNOID_EXPORT int cnoid::deleteUnmanagedRTCs()
     }
 
 #ifdef OPENRTM_VERSION110
-    for(int i=0; i < rtcs.size(); ++i){
+    for(size_t i=0; i < rtcs.size(); ++i){
         RTC::RTObject_impl* rtc = rtcs[i];
         if(managedComponents.find(rtc) == managedComponents.end()){
             RTC::ExecutionContextList_var eclist = rtc->get_participating_contexts();
@@ -363,7 +363,7 @@ CNOID_EXPORT int cnoid::deleteUnmanagedRTCs()
             if(myEClist->length() > 0 && !CORBA::is_nil(myEClist[0])){
             	OpenRTM::ExtTrigExecutionContextService_var myEC = OpenRTM::ExtTrigExecutionContextService::_narrow(myEClist[0]);
                 RTC::RTCList rtcs = myEC->get_profile()->participants;
-            	for(int i=0; i < rtcs.length(); ++i){
+            	for(size_t i=0; i < rtcs.length(); ++i){
             		myEC->remove_component(rtcs[i]);
             	}
             }
@@ -400,7 +400,7 @@ CNOID_EXPORT int cnoid::deleteUnmanagedRTCs()
     }
 #endif
 
-    for(int i=0; i < rtcs.size(); ++i){
+    for(size_t i=0; i < rtcs.size(); ++i){
         RTC::RTObject_impl* rtc = rtcs[i];
         if(managedComponents.find(rtc) == managedComponents.end()){
         	//std::cout << rtc->getInstanceName() << std::endl;
@@ -433,7 +433,7 @@ CNOID_EXPORT int cnoid::deleteUnmanagedRTCs()
 #ifdef OPENRTM_VERSION110
     		OpenRTM::ExtTrigExecutionContextService_var myEC = OpenRTM::ExtTrigExecutionContextService::_narrow(myEClist[0]);
                 RTC::RTCList rtcs = myEC->get_profile()->participants;
-    		for(int i=0; i < rtcs.length(); ++i){
+    		for(size_t i=0; i < rtcs.length(); ++i){
     			myEC->remove_component(rtcs[i]);
     		}
 #else
